pyproject.toml
[project]
name = "receitas-api2"
version = "0.1.0"
description = "API de receitas"
readme = "README.md"
requires-python = ">=3.13"

dependencies = [
    "alembic==1.17.2",
    "annotated-doc==0.0.4",
    "annotated-types==0.7.0",
    "anyio==4.12.0",
    "certifi==2025.11.12",
    "click==8.3.1",
    "colorama==0.4.6",
    "dnspython==2.8.0",
    "email-validator==2.3.0",
    "fastapi==0.123.0",
    "fastar==0.8.0",
    "greenlet==3.2.4",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httptools==0.7.1",
    "httpx==0.28.1",
    "idna==3.11",
    "itsdangerous==2.2.0",
    "Jinja2==3.1.6",
    "Mako==1.3.10",
    "markdown-it-py==4.0.0",
    "MarkupSafe==3.0.3",
    "mdurl==0.1.2",
    "orjson==3.11.4",
    "psycopg2-binary==2.9.11",
    "pydantic==2.12.5",
    "pydantic-extra-types==2.10.6",
    "pydantic-settings==2.12.0",
    "pydantic_core==2.41.5",
    "Pygments==2.19.2",
    "python-dotenv==1.2.1",
    "python-multipart==0.0.20",
    "PyYAML==6.0.3",
    "rich==14.2.0",
    "rich-toolkit==0.17.0",
    "rignore==0.7.6",
    "sentry-sdk==2.46.0",
    "shellingham==1.5.4",
    "SQLAlchemy==2.0.44",
    "sqlmodel==0.0.27",
    "starlette==0.50.0",
    "typer==0.20.0",
    "typing-inspection==0.4.2",
    "typing_extensions==4.15.0",
    "ujson==5.11.0",
    "urllib3==2.5.0",
    "uvicorn==0.38.0",
    "watchfiles==1.1.1",
    "websockets==15.0.1"
]

[tool.uv]
# coloque aqui dependências de desenvolvimento (motor de migração, watch, CLI de dev etc.)
dev-dependencies = [
    "fastapi-cli==0.0.16",
    "fastapi-cloud-cli==0.5.2"
]